Add files via upload
Solution of the problem 'Find the first node of the loop in linked list'
Given a singly linked list of N nodes. Find the first node of the loop if the linked list has a loop. If a loop is present return the node data of the first node of the loop else return -1.