Add files via upload
Solution of problem 'Element with left side smaller and right side greater'
Given an unsorted array of size, N. Find the first element in an array such that all of its left elements are smaller and all right elements to it are greater than it.
Note: Left and right side elements can be equal to the required element. And extreme elements cannot be required elements.