Add files via upload
Solution of problem 'Alternate positive and negative numbers'.
Given an unsorted array Arr of N positive and negative numbers. Your task is to create an array of alternately positive and negative numbers without changing the relative order of positive and negative numbers.
Note: The array should start with a positive number.