Add files via upload
Solution the problem 'Convert an array to reduced form'
Given an array with N distinct elements, convert the given array to a reduced form where all elements are in the range from 0 to N-1. The order of elements is the same, i.e., 0 is placed in the place of the smallest element, 1 is placed for the second smallest element, and N-1 is placed for the largest element.
Note: You don't have to return anything. You just have to change the given array.